requirements.txt
django==2.2.1
celery==4.4.6
redis==3.4.1
django-cors-headers==3.0.2
django-filter==2.1.0
django-compat==1.0.15
djangorestframework==3.9.4
django-sslserver==0.20
psycopg2-binary>=2.8
requests==2.24.0
beautifulsoup4==4.9.1
coverage==3.6
